+++ app/forms/transfer/task/form_m_task_form.rb
@@ -1,0 +1,7 @@
+class Transfer::Task::FormMTaskForm < BaseTaskForm
+  attribute :received_form_m
+  attribute :received_title_plans
+  attribute :cleared
+  attribute :signed
+  attribute :saved
+end